Initial commmit. Move SHIPcal 1.1 to legacy folder. Start weather class.
