+++ .gitignore
@@ -137,3 +137,5 @@ settings.py
 
 .vscode/
 # End of https://www.gitignore.io/api/python
+
+ressspi_solatom/
