+++ README.md
@@ -11,7 +11,7 @@ SHIPcal is an open-source code to simulate solar heat applications for industria
 #Resuming commits and contributions begining October 2018
 
 #Improvements from 2019
--Juan Antonio Aramburo Pasapera (CIMAV) Changed demand creator funtion to improve the accuracy of the results
+-Juan Antonio Aramburo Pasapera (CIMAV) Changed demand creator function to improve the accuracy of the results
 -Juan Antonio Aramburo Pasapera (CIMAV) Changed fixed raw water temperature function. Now it depends on the TMY
 - Miguel Frasquet Herraiz (SOLATOM) Added new reg to allow external front-ends call ressspi (see an example of frontend at https://github.com/mfrasquet/frontendRessspi
