+++ elements.py
@@ -1,0 +1,35 @@
+class Element:
+    """
+    Base class inherits every other element in SHIPcal.
+    """
+    
+    # Primary
+    
+    P_p_in = 0
+    m_dot_p_in = 0
+    T_p_in = 0
+    h_p_in = 0
+    s_p_in = 0
+    
+    P_p_out = 0
+    m_dot_p_out = 0
+    T_p_out = 0
+    h_p_out = 0
+    s_p_out = 0
+    
+    # Secondary
+    
+    P_s_in = 0
+    m_dot_s_in = 0
+    T_s_in = 0
+    h_s_in = 0
+    s_s_in = 0
+    
+    P_s_out = 0
+    m_dot_s_out = 0
+    T_s_out = 0
+    h_s_out = 0
+    s_s_out = 0
+    
+    # Thermal innertia
+
