+++ legacy/General_modules/Consumo_random_year.py
