+++ legacy/Meteo_modules/Bakersfield.dat
