+++ legacy/Solar_modules/EQSolares.py
