+++ legacy/Solar_modules/char_day.py
