+++ legacy/sliced_SHIPcal.py
