+++ tests/test_weather.py
@@ -1,0 +1,4 @@
+from ..weather import Weather
+
+celaya_loc_tmy = "/home/jaarpa/aProjects/solatom/SHIPcal/tests/Celaya.dat"
+celaya = Weather(celaya_loc_tmy)
